update test script to ignore missing go.mod
As the original technoweenie/grohl package predates the introduction
of Go modules in Go 1.11 and, in particular, the change to the
default value of the GO111MODULE setting to "true" in Go 1.16, we
need to explicitly set GO111MODULE=off in order to continue to
test this package without a go.mod file.

We also add an explicit shebang to match the script/fmt script.