always use new byte array for grohl stack traces
As we are updating the grohl package to meet modern Go linter
requirements, one issue is that the Stack() method currently
uses a shared pool of 1 MB byte slices, and the staticcheck
linter reports that sync.Pool should only be used with pointers,
not structs such as slices.

This in turn points to another potential issue with the existing
code, which is that it returns a slice referencing the written
portion of the byte slice is retrieved from the pool, but does so
after returning the slice to the pool, which means it is available
to be overwritten by another call to Stack() since the same byte
slice could be re-issued from the pool, and slices share their
underlying data arrays.

To resolve both issues, we simply remove the use of the shared
pool and allocate a fresh 64 kB array for each call to Stack().
The smaller size should be ample for a stack trace, and this
avoids the need to "freeze" the written set of data in a shared
array until it is utilized by a caller of the Stack() function.