add golangci-lint checks
We add golangci-lint to the existing gofmt checks we run in our
test script.  For now we disable three of the default linters,
errcheck, gosimple, and staticcheck, as we will deal with some of
the issues they report in subsequent commits.  (We do not anticipate
resolving the issues reported by the errcheck linter as they would
require changes to the existing API.)