always double-quote log key values for Splunk
The legacy behaviour of the grohl module's log formatting code is to
check if a string contains a space, and if so, quote it either in
single quotes or double quotes, depending on whether the string
contains either of both of those characters.  In the particular case
where a string contains both spaces and double quotes, but not single
quotes, it is single-quoted, with the double quotes escaped using
backslashes.

The single-quoted format is not well parsed by Splunk, which understands
each space-separated part of the string as a separate field unless the
key's value is double-quoted, and so only the first word of the value
string (plus the opening single quote) is logged for the key.

To resolve this, we simply always emit string values in double quotes.
We retain the legacy escaping of backslash characters with another
backslash (although this may not be necessary with Splunk), but we
convert double-quote characters into single quotes instead of trying
to escape them, which should always be safe with Splunk.